feat: physics space prioritization
Spaces can be given a priority with `PhysicsServer2/3D::space_set_priority(RID p_space, int p_priority)` which will cause higher priority spaces to be processed first in `flush_queries()` and `step()`.
This predictable ordering of execution allows for spaces to be "nested" if a scene tree has a child node in a different space from its parent while still retaining the global transform of the parent (i.e. a "spaceship" rigidbody is the parent of a "player" rigidbody, with the player put into a different space while inside the spaceship)